Actor Wall SFML init

+++ src/ui/menu.hpp
@@ -10,6 +10,7 @@ namespace sokoban
     {
         using namespace sokoban::ui::gui;
         using namespace sokoban::util;
+
         class Menu
         {
         public:
@@ -17,14 +18,13 @@ namespace sokoban
 
             virtual ~Menu();
 
-            int launch_application() const;
+            unsigned short launch_application() const;
 
         private:
-            MainFrame *main_frame;
-            Logger *logger{};
+            MainFrame* main_frame;
+            Logger* logger{};
         };
     }
 }
-
 
 #endif //SOKOBAN_MENU_HPP
